homework-3 ревью: перенести валидаторы в отдельный файл validation.ts
